Renamed from Teemu's org.* packages to non-teemu com.vaadin .
